restructure navigator
